% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Function server to run TradeOi app}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{file in}

\item{output}{file create}

\item{session}{for reactive and observe}
}
\description{
Some options are programmed
Suppress missing test error
Maximum importable file size
Feature in parallel
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
