% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_error.R
\name{are_error_in_DS}
\alias{are_error_in_DS}
\title{Function to check if there is an error in dataset 1}
\usage{
are_error_in_DS(log1_dataset)
}
\arguments{
\item{log1_dataset}{dataset with error info}
}
\description{
Function to check if there is an error in dataset 1
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
