% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{are_error_in_DS1}
\alias{are_error_in_DS1}
\title{Function to check if there is no empty table during all the features}
\usage{
are_error_in_DS1(log1_dataset)
}
\arguments{
\item{log1_dataset}{dataset with error info}
}
\description{
Function to check if there is no empty table during all the features
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
