% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_names.R
\name{change_names1}
\alias{change_names1}
\title{Function to check and change the names}
\usage{
change_names1(
  input_dataset,
  product_table,
  var_productUI,
  var_partnerUI,
  var_reporter,
  var_partner,
  var_product,
  var_value,
  var_quantity,
  var_quantity_unit,
  var_time,
  var_product_PDT,
  var_revision_PDT,
  date_range,
  input_error
)
}
\arguments{
\item{input_dataset}{last dataset}

\item{product_table}{product dataset}

\item{var_productUI}{variable}

\item{var_partnerUI}{variable}

\item{var_reporter}{variable for reporter}

\item{var_partner}{variable for partner}

\item{var_product}{variable for product}

\item{var_value}{variable for value}

\item{var_quantity}{variable for quantity}

\item{var_quantity_unit}{variable for quantity unit}

\item{var_time}{variable for time}

\item{var_product_PDT}{variable for product in product table}

\item{var_revision_PDT}{variable for revision in product table}

\item{date_range}{the chosen date}

\item{input_error}{error dataset}
}
\description{
Function to check and change the names
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
