% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_end.R
\name{date_end1}
\alias{date_end1}
\title{Function generating the date and time of the start of calculations}
\usage{
date_end1(log_prev)
}
\arguments{
\item{log_prev}{last hour}
}
\description{
Function generating the date and time of the start of calculations
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
