% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublons.R
\name{doublons1}
\alias{doublons1}
\title{Function to check and delete doublons}
\usage{
doublons1(input_dataset, empty_parameter)
}
\arguments{
\item{input_dataset}{last dataset}

\item{empty_parameter}{empty data}
}
\description{
Function to check and delete doublons
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
