% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{heterogeneity1}
\alias{heterogeneity1}
\title{Function to detect heterogeneity in the time series}
\usage{
heterogeneity1(input_dataset, empty_parameter, checkbox2, radio_outliers)
}
\arguments{
\item{input_dataset}{last dataset}

\item{empty_parameter}{empty data}

\item{checkbox2}{compute "delete heterogenious time series}

\item{radio_outliers}{define the type of computation for outliers}
}
\description{
Function to detect heterogeneity in the time series
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
