% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_output.R
\name{is_empty_output1}
\alias{is_empty_output1}
\title{Function to check if there is no empty table during all the features}
\usage{
is_empty_output1(log_dataset, note_txt, IEx, IEx_1, error_input)
}
\arguments{
\item{log_dataset}{dataset with error info}

\item{note_txt}{info}

\item{IEx}{}

\item{IEx_1}{}

\item{error_input}{error info}
}
\description{
Function to check if there is no empty table during all the features
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
