% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_parameters.R
\name{is_empty_parameter1}
\alias{is_empty_parameter1}
\title{Function to check if there is no empty parameters}
\usage{
is_empty_parameter1(input_dataset)
}
\arguments{
\item{input_dataset}{dataset}
}
\description{
Function to check if there is no empty parameters
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
