% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_heterogeneous.R
\name{non_heterogeneous1}
\alias{non_heterogeneous1}
\title{Function to keep only heterogeneous series}
\usage{
non_heterogeneous1(input_dataset, empty_parameter)
}
\arguments{
\item{input_dataset}{last dataset}

\item{empty_parameter}{empty data}
}
\description{
Function to keep only heterogeneous series
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
