% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_HS.R
\name{non_HS1}
\alias{non_HS1}
\title{Function to verify if the table revision and the data set are compatibles}
\usage{
non_HS1(input_dataset, out_verif_table, log_table, empty_parameter)
}
\arguments{
\item{input_dataset}{last dataset}

\item{out_verif_table}{table clear}

\item{log_table}{table with error}

\item{empty_parameter}{empty data}
}
\description{
Function to verify if the table revision and the data set are compatibles
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
