% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_detection.R
\name{outliers_detection1}
\alias{outliers_detection1}
\title{Function to detect outlier with robust regression and Z-modify}
\usage{
outliers_detection1(input_dataset, var_time, empty_parameter, radio_outliers)
}
\arguments{
\item{input_dataset}{last dataset}

\item{var_time}{variable for the time}

\item{empty_parameter}{empty data}

\item{radio_outliers}{define the type of computation for outliers}
}
\description{
Function to detect outlier with robust regression and Z-modify
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
