% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_indices.R
\name{prepare_indices1}
\alias{prepare_indices1}
\title{Function to prepare the data for feature indice}
\usage{
prepare_indices1(
  input_dataset,
  time_break,
  indices_bc,
  date_to_date,
  date_range,
  empty_parameter,
  checkbox3
)
}
\arguments{
\item{input_dataset}{last dataset}

\item{time_break}{the chosen date}

\item{indices_bc}{indices chosen}

\item{date_to_date}{date to date or not}

\item{date_range}{month/year/quarter}

\item{empty_parameter}{empty data}

\item{checkbox3}{compute "indice"}
}
\description{
Function to prepare the data for feature indice
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
