% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product.R
\docType{data}
\name{product}
\alias{product}
\title{Product table from ITC}
\format{
A data frame with 7668 rows and 3 variables:
\describe{
\item{PRODUCT_CD}{Product code}
\item{LABEL_EN}{Description of product code}
\item{FG_REVISION}{Time for revision}
...
}
}
\source{
\url{ITC/DMD/TMI}
}
\usage{
product
}
\description{
Product table from ITC
}
\keyword{datasets}
