% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision.R
\docType{data}
\name{revision}
\alias{revision}
\title{Revision table from ITC}
\format{
A data frame with 2848 rows and 3 variables:
\describe{
\item{REPORTER_CD}{Reporter country code}
\item{year}{Year}
\item{class}{Type of class}
...
}
}
\source{
\url{ITC/DMD/TMI}
}
\usage{
revision
}
\description{
Revision table from ITC
}
\keyword{datasets}
