% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_app.R
\name{run_app}
\alias{run_app}
\title{Function to launch TradeOI app}
\usage{
run_app(options = list())
}
\arguments{
\item{ui}{ui file}

\item{server}{server file}
}
\description{
Function to launch TradeOI app
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
