% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strKey.R
\name{strKey}
\alias{strKey}
\title{Function to generate the primary key/vector}
\usage{
strKey(reporter, product, partner, var_key_add)
}
\arguments{
\item{reporter}{variable reporter}

\item{product}{variable product}

\item{partner}{variable partner}

\item{var_key_add}{variable add}
}
\description{
Function to generate the primary key/vector
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
