% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_parameters.R
\name{test_parameters1}
\alias{test_parameters1}
\title{Function to verify and create some errors during all the procedure}
\usage{
test_parameters1(
  input_dataset,
  product_table,
  revision_table,
  var_productUI,
  var_partnerUI,
  var_reporter,
  var_partner,
  var_product,
  var_value,
  var_quantity,
  var_quantity_unit,
  var_time,
  var_product_PDT,
  var_revision_PDT,
  var_key_add,
  time_breakdown,
  date_range,
  indices_type,
  agr_partner,
  year_to_date,
  checkbox1,
  checkbox2,
  checkbox3
)
}
\arguments{
\item{input_dataset}{laste dataset}

\item{product_table}{table product import}

\item{revision_table}{table revsion import}

\item{var_productUI}{name of variable product}

\item{var_partnerUI}{name of variable partner}

\item{var_reporter}{name of variable reporter}

\item{var_partner}{name of variable partner}

\item{var_product}{name of variable product}

\item{var_value}{name of variable value}

\item{var_quantity}{name of variable quantity}

\item{var_quantity_unit}{name of variable quantity unit}

\item{var_time}{name of variable time}

\item{var_product_PDT}{name of variable product in the product table}

\item{var_revision_PDT}{name of variable revision in the revision table}

\item{var_key_add}{variable add}

\item{time_breakdown}{the chosen date (end)}

\item{date_range}{the chosen date}

\item{indices_type}{indices choose}

\item{agr_partner}{aggregate or not on partner}

\item{year_to_date}{year to date or not}

\item{checkbox1}{compute "unstable HS revision}

\item{checkbox2}{compute "delete heterogenious time series}

\item{checkbox3}{compute "indice"}
}
\description{
Function to verify and create some errors during all the procedure
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
