% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_code.R
\name{time_code1}
\alias{time_code1}
\title{Function to modify and standardize the time variable}
\usage{
time_code1(input_dataset, pct_missdata, empty_parameter, checkbox1)
}
\arguments{
\item{input_dataset}{last dataset}

\item{pct_missdata}{\% of chosen missing data}

\item{empty_parameter}{empty data}

\item{checkbox1}{compute "unstable HS revision}
}
\description{
Function to modify and standardize the time variable
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
