% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\docType{data}
\name{trade}
\alias{trade}
\title{Trade table from ITC}
\format{
A data frame with 20334 rows and 8 variables:
\describe{
\item{reporter}{Reporter country code}
\item{flow}{Description of trade flow (Import or Export)}
\item{month}{Variable for the time}
\item{ntlc}{Product code}
\item{parttr}{Partner code}
\item{qty1}{Quantity}
\item{unit1}{Quantity unit}
\item{val}{Value}
...
}
}
\source{
\url{ITC/DMD/TMI}
}
\usage{
trade
}
\description{
Trade table from ITC
}
\keyword{datasets}
