% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_quant.R
\name{unit_quant1}
\alias{unit_quant1}
\title{Function to check and change the unit quantity}
\usage{
unit_quant1(input_dataset, empty_parameter)
}
\arguments{
\item{input_dataset}{last dataset}

\item{empty_parameter}{empty data}
}
\description{
Function to check and change the unit quantity
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
