% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_value_indices.R
\name{unit_value_indices1}
\alias{unit_value_indices1}
\title{Function to feature indices}
\usage{
unit_value_indices1(
  input_dataset,
  product_level,
  var_reporter,
  var_partner,
  date_to_date,
  empty_parameter,
  time_break,
  checkbox3
)
}
\arguments{
\item{input_dataset}{last dataset}

\item{product_level}{level for product code}

\item{var_reporter}{variable reporter country}

\item{var_partner}{variable partner country}

\item{date_to_date}{if date to date is check}

\item{empty_parameter}{data empty}

\item{time_break}{the chosen date}

\item{checkbox3}{compute "indice"}
}
\description{
Function to feature indices
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
