% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verif_HS.R
\name{verif_HS1}
\alias{verif_HS1}
\title{Function to standardize HS code}
\usage{
verif_HS1(input_dataset, product_table, empty_parameter)
}
\arguments{
\item{input_dataset}{dataset to produce}

\item{product_table}{the product table than your import}

\item{empty_parameter, }{empty file}
}
\description{
Function to standardize HS code
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
