% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chang_rev.R
\name{chang_rev1}
\alias{chang_rev1}
\title{Function to check and change the revision}
\usage{
chang_rev1(input_dataset, table_revision, empty_parameter)
}
\arguments{
\item{input_dataset}{last dataset}

\item{table_revision}{revision dataset}

\item{empty_parameter}{empty dataset}
}
\description{
Function to check and change the revision
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
