% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covert_rate.R
\name{covert_rate1}
\alias{covert_rate1}
\title{Function to create the covert rate}
\usage{
covert_rate1(
  original_ds,
  non_HS_ds,
  doublons_ds,
  qty_unit_ds,
  chmt_rev_ds,
  incompl_series_ds,
  outliers_ds,
  non_homogeneous_ds,
  indices_ds,
  detected_outliers,
  series_hetero,
  empty_parameter
)
}
\arguments{
\item{original_ds}{original dataset}

\item{non_HS_ds}{non HS dataset}

\item{doublons_ds}{doublon dataset}

\item{qty_unit_ds}{quantity unit dataset}

\item{chmt_rev_ds}{revision change dataset}

\item{incompl_series_ds}{incomplete series dataset}

\item{outliers_ds}{outliers dataset}

\item{non_homogeneous_ds}{non homogeneous dataset}

\item{indices_ds}{indices dataset}

\item{detected_outliers}{outliers detected dataset}

\item{series_hetero}{series hetero dataset}

\item{empty_parameter}{empty data}
}
\description{
Function to create the covert rate
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
