% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_outliers.R
\name{out_outliers1}
\alias{out_outliers1}
\title{Function to delete the outliers}
\usage{
out_outliers1(
  input_dataset_outliers,
  input_dataset_hetero,
  input_dataset_outliers2,
  empty_parameter,
  radio_outliers
)
}
\arguments{
\item{input_dataset_outliers}{dataset without outliers}

\item{input_dataset_hetero}{dataset with heterogeneity}

\item{input_dataset_outliers2}{dataset with outliers}

\item{empty_parameter}{empty data}

\item{radio_outliers}{define the type of computation for outliers}
}
\description{
Function to delete the outliers
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
