% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{replacement1}
\alias{replacement1}
\title{Function to replace or delete the outliers}
\usage{
replacement1(input_dataset, empty_parameter)
}
\arguments{
\item{input_dataset}{last dataset}

\item{empty_parameter}{empty data}
}
\description{
Function to replace or delete the outliers
}
\author{
Yann Rivallant \href{mailto:yrivallant@intracen.org}{yrivallant@intracen.org}
}
